chores: renamed push-notification utils to local-notification
